migrate script
We needed to migrate the old liked quotes to a new system that handled
when they
were liked. So this is what we used. There is a pausecomp function that
we used so that they are not all liked at exactly the same time.

Hopefully this all works as expected.