Testing email and putting in server latency simulation
Simulating latency to properly see the progress bar in action. You'll
notice it is a random delay in the publish functions. We added a global
function file and a global Javascript function to get a random number
between two numbers.

Also I added the email package because the next step is to build a
submit quote thing that just emails us the quote. Build things that
don't scale remember.