Redefining the off canvas menu
After some extensive research into a few different ways off doing this,
the solution presented itself, quite simply. This is Meteor, so we can
do it quite differently. The menu will be its own page and template and
route. This will be quick enough due to the live nature of the Meteor
app.

We will see how we go anyway. Maybe research the Esc key to go back to
previous page after menu or something. Or probably just put a little X
in the corner. Here we are. I think we are on the right track here.