Beginning are taking shape. Random searches now working.
Probably we will do this a bit differently, but you get the basic idea.

Now, next we will have it instead of searching the quotation text, we
will have it so that it searches a "tag" field or something like that. a
"words" field probably with a json array in it..... probably limit the
number of words maybe.